Relocate stocktake and stocktake line tests closer to implementation
