+++ service/src/stocktake_line/mod.rs
@@ -15,10 +15,6 @@ pub mod insert;
 pub mod query;
 pub mod update;
 pub mod validate;
-
-#[cfg(test)]
-mod tests;
-
 pub trait StocktakeLineServiceTrait: Sync + Send {
     fn get_stocktake_lines(
         &self,
